Included FETCH_PLANS action in SiteStore
This is first step for implementing a feature which allows user to associate a custom domain(https://github.com/wordpress-mobile/WordPress-Android/issues/7923).
The user can associate a custom domain only if he/she has a 'domain credit' in his current plan.
This change introduces fetch plans action into SiteStore from which the application can get all the plans of the user and then take a decision on whether he has a 'domain credit'.