Adds fetchList function to ListItemDataSource for easier reusability.
In the previous iteration, `ListStore` was responsible for deciding how to fetch a list for the given type. That meant that, to add a new type, `ListStore` had to be modified. Furthermore, it made it very hard to work within a different module and still take advantage of `ListStore`. To avoid this issue, a new function has been added to `ListItemDataSource` which tells the `ListStore` how to fetch a list given a `ListDescriptor`, whether to load more data and the offset which will be provided by `ListStore`.

Note that this is just the initial stab at fixing this issue. I expect some further refactors for this. There are two things I don't like with this approach:

1. If it's possible, I don't want the clients to tell how to fetch a list or an item. This was something I wanted to fix for `fetchItem` before, now we'll also have to do it for `fetchList`.
2. We are passing the fetch function around too many times. Hopefully fixing the first issue will also fix this one as well.