Assign NotificationModel siteId from individual notification response
Fixed an issue where the siteId was erroneously being set from an external siteModel and not using the api response for the individual notification. This required updating all the tests and most methods as the siteModel argument was no longer relevant since siteId is not a parameter to filter by when fetching notifications from the api. Updated tests as well.