Added the usernameCanBeChanged column to the AccountModel table and updated the relevant tests to verify it's addition.
Updated all the respective setters, equals, copy, etc functions that reference it.
Modified the AccountModelTest's testCopyOnlyAccountSettingsAttributes to include the new property.
Modified the AccountSqlUtilsTest to verify that the value is being persisted correctly.
modified the WellSqlConfig’s onUpgrade so that the AccountModel table can be altered with the new property and bumped the database version