Feature: Add post date gmt to post model
This 'post_date_gmt' value from the post response is actually the value
that is going to be needed on the UI layer to at least show for when a
scheduled post is due. As such, this is the main date that is going to
be needed.

However, in addition to that 'post_date_gmt', I decided to keep the
'post_modified_gmt' as well, use the 'modified' post model field for it,
since it might end up being needed after all. But, mind the fact that it
is not needed on the UI atm, the 'post_date_gmt' is enough as it is the
only date to be used, and that will only applicable to scheduled posts.