Update build.gradle
A work around to avoid
`XML RPC Symbol not found` the crops up when using a Composite build with FluxC due to sqlite-jdbc library used in Room doesn't work on Mac.

It is fixed Room ~ v2.4.0.  Till FluxC is upgraded to it.  This would prevent the issue for anyone going to use a M1 Mac (as more people are ordering now)