Build: (GE) Enable capturing of task input files for build comparison
Context: Build scans capture hashes of task inputs, to enable
identifying changes to inputs when comparing builds, among other
features. By default, an overall hash value for each task input property
is captured. This enables identifying which properties changed for a
task (e.g. the source or the classpath for Java compilation), but not
which individual files changed. In order to identify file changes, the
paths and content hashes of each individual input file must be captured,
which can be enabled.