Build: (GE) Disable build scan background upload for all CI builds
Context: By default, build scans are uploaded in the background after
the build has finished. This allows the build to finish sooner, but can
be problematic in build environments (e.g. ephemeral CI agents) that
terminate as soon as the build is finished, as the upload may be
terminated before it completes. Background uploading should be disabled
for such environments.