Build: Add detekt to fluxc
This commit adds Detekt version '1.21.0' to this repo.

In addition to that, this commit configures Detekt basing the
configuration upon the default config and overriding specific rules via
the 'detekt.yml' file (like 'GlobalCoroutineUsage'). This is done in
order to make it easier to update Detekt going forward instead of
depending on running the 'detektGenerateConfig' command to update the
'detekt.yml' file and then fine-tuning the rules.

Apart from that, the 'detekt-formatting' plugin was added to this
configuration in order to have the 'Formatting' section being reported,
alongisde all its corresponding formatting violations. However, the
'autoCorrect' configuration flag was set to 'false' in order to not have
the codebase being automatically reformatted whenever Detekt gets
triggred.