Build: Add detekt baseline to fluxc
This Detekt baseline suppresses the existing issues (more than 1K in
total) to have Detekt running as if the codebase has zero issues. This
helps to start utilising Detekt immediately without having to worry
about the existing issues.

Any new issues that will be added to the codebase right after will be
reported and the build will fail accordingly.

------------------------------------------------------------------------

Note that this baseline is split into 3 parts:
1) <!-- fluxc -->
2) <!-- plugins:woocommerce -->
3) <!-- example -->

To create this baseline, and because Detekt baseline is not working as
expected for multi-module projects, the below commands were first used
individually, per module, that is, instead of just issuing this single
'./gradlew detektBaseline' command:
1) ./gradlew fluxc:detektBaseline
2) ./gradlew plugins:woocommerce:detektBaseline
3) ./gradlew example:detektBaseline

Then, the each result were copy-pasted into this single 'baseline.xml',
and the '<!-- module -->' group comments were appended afterwards.