Analysis: Suppress magic number warning for wc order list activity
The 'ArrayList' initial capacity was explicitly set to a specific number
depending on the amount of parameters, per API call. However, this is
totally optional and not affecting the outcome of the function, nor its
performance. The default initial capacity is 10 anyway and as such,
having something less than (3, 4 or 5 in this case), even slightly more
than that (ie 20, 50 or 100), doesn't help much, performance-wise.

Description: "This rule detects and reports usages of magic numbers in
the code. Prefer defining constants with clear names describing what the
magic number means."

For more info see: https://detekt.dev/docs/rules/style/#magicnumber