Analysis: Exclude kotlinx android synthetic imports for wildcard import
With this configuration all 'kotlinx.android.synthetic' related imports
will end up being ignored by the 'WildcardImport' rule. This is done
because the 'Kotlin Android Extensions' plugin is still being utilized
within the 'example' module. When this plugin will get replaced with
'View Binding', this rule configuration should be reverted as it will no
longer be applicable for this repo.

This change removes '49' warnings from the baseline.