Analysis: Suppress too generic exception caught warnings for example mod
Description: "This rule reports catch blocks for exceptions that have a
type that is too generic. It should be preferred to catch specific
exceptions to the case that is currently handled. If the scope of the
caught exception is too broad it can lead to unintended exceptions being
caught."

For more info see: https://detekt.dev/docs/rules/exceptions
#toogenericexceptioncaught