Build: Extract common min sdk version to settings for all modules
This will now make sure that from now on all module will be pointing to
the same 'minSdkVersion' versions, making such upgrades much more
straightforward as that will apply to all modules. This will alleviate
the risk of forgetting to update a specific module, or set of modules
due to an oversight.