Deps: Remove unnecessary org apache commons lang dependency for woo
This dependency was added as part of the
445601941d74d1430287e56b76596244f567e469 commit on the
'plugins:woocommerce' module, most probably in order to be later used as
part of the d0b92ceab8aee19de4e2148d40385d49dad20ade commit, but this
time, on the 'example' module on 'WCOrderListActivity' class in order to
import 'org.apache.commons.lang3.time.DateUtils'.

However, the 'example' module already has such a
'org.apache.commons:commons-lang3' dependency declaration. As such, this
similar dependency declaration on the 'plugins:woocommerce' module is
totally unused and can be safely removed.