Analysis: Remove unnecessary commons lang random string utils
The 'RandomStringUtils.randomAlphanumeric/Alphabetic/Number(...)' can be
safely replaced with a Kotlin equivalent solution.

As such, this commit also removes the unnecessary org apache commons
lang Android test dependency for the 'example' module.

FYI: Since this is a random string utility, it is difficult to unit
test that. However, I did manually unit test those for correctness,
compared those to the Apache Commons Lang3 'RandomStringUtils' and
verified that the results are the expected ones. Also, the connected
tests can verify correctness of this change as well.