wip - PLEASE REBASE THIS COMMIT BEFORE MERGING
I'm pushing this work in progress at the start of my leave. This commit
has some work to fetch the block-based-theme flag, but needs some more
work. With this it should be possbible to wire up conditional UI for the
pages list to insert an entry-point for a webview flow of the editor.

The goal is to optimistically fetch the theme info any time the site
changes, and also once when the page list is shown. This way, we can
hope to have the most up-to-date value for the flag.