add temporary support for prompts v2 endpoint
The idea is supporting v2 and v3 at the same time so the client can toggle
between the two implementations remotely when we decide to flip the switch, to
make sure web and mobile will start seeing the same thing at the same time.

This is needed because the daily prompts are not aligned between v2 and v3,
meaning that shipping the v3 changes in only 1 platform will cause a mismatch
in what the user sees in different platforms (if they use mobile and web, for
instance).