Build: Update kotlin to 1.6.20
Release Notes: https://github.com/JetBrains/kotlin/releases/tag/v1.6.20

------------------------------------------------------------------------

This is a tmp solution which will be removed once the 'Catalog' related
https://github.com/Automattic/android-dependency-catalog/pull/20 PR
gets ready for review and the CI is able to publish the new catalog.

------------------------------------------------------------------------

FYI: Instead of updating Kotlin to this newer '1.6.20' patch version, a
better alternative would have been to replace the deprecated
'kotlin-android-extensions' plugin with 'kotlin-parcelize'. However,
doing such a migration is not trivial and will require a lot of work.
For instance, one could search for all the 'kotlinx.android.synthetic'
and 'kotlinx.android.parcel' related imports. Those would need to be
replaced with 'ViewBinding' and 'kotlinx.parcelize' respectively.

As a reference see WPAndroid such issues and their multiple PRs:
- Replace Synthetic Accessors with ViewBinding #14845
https://github.com/wordpress-mobile/WordPress-Android/issues/14845
- ViewBinding: Remove Kotlin Android Extensions (Replace with Kotlin
Parcelize) #14902
https://github.com/wordpress-mobile/WordPress-Android/pull/14902