Deps: Update wellsql to 1.9.0 (via catalog's pr-sha combination)
FYI: This change was done for testing purposes and until the below
'Android Dependency Catalog' #25 PR gets merged to 'trunk'. When that's
done, the 'catalogVersion' will be updated to '1.13.0' instead.

------------------------------------------------------------------------

Since this update is a jump from '1.11.0' to '1.13.0', this mean that
the below '1.12.0' updates are incorporated in this change:
- 'androidGradlePlugin' -> '8.0.2'
- 'kotlin'              -> '1.6.20'
- 'restAssured'         -> '5.3.1'

Catalog PR ('AGP + Kotlin'): [Build] Upgrade Gradle to 8.2.1 and
AGP to 8.0.2 #20
- https://github.com/Automattic/android-dependency-catalog/pull/20
Catalog PR ('restAssured'): [Deps] Update restAssured to 5.3.1 #21
- https://github.com/Automattic/android-dependency-catalog/pull/21

------------------------------------------------------------------------

Also, the 'androidSourcesJar' publishing configuration got removed,
since this pr-sha combination is also pointing to the newer version of
the 'automattic-publishToS3' plugin, version '0.8.0':

Release Notes: https://github.com/Automattic/publish-to-s3-gradle-plugin
/releases/tag/0.8.0

Plugin PR: Adds support for Android publishing block #31
- https://github.com/Automattic/publish-to-s3-gradle-plugin/pull/31
Plugin Release:
Catalog PR: Update publish-to-s3 Gradle plugin to 0.8.0 #24
- https://github.com/Automattic/android-dependency-catalog/pull/24