Remove Migration and fallback to destructive migration
Since UPSERT is not supported by all Android API versions and the actual data
inside the Prompts table changed along with the schema, it will be best to avoid
the effort of maintaing stale data and just rely on Room's fallback to
destructive migration, which will drop the table (and data) and create a new one
with the correct schema.

This shouldn't have many downsides as we fetch new data as soon as the app opens
and the user reaches the Dashboard. The only possible caveat is for users that
have reminders setup with Prompts enabled, that might see a notification without
Blogging Prompts if they don't open the app after updating it.