Build: Add dependency analysis plugin (via catalog's pr-sha combination)
FYI: This change was done for testing purposes and until the below
'Android Dependency Catalog' #32 PR gets merged to 'trunk'. When that's
done, the 'catalogVersion' will be updated to '1.19.0' instead.

------------------------------------------------------------------------

Android Dependency Catalog PR: [Migrate] [Dependency Analysis] Add
Android Gradle Plugin #31
- https://github.com/Automattic/android-dependency-catalog/pull/31

------------------------------------------------------------------------

FYI: This commit adds a previous version of the plugin (1.28.0) and not
the latest version (1.32.0). This is because the latest version gets
stuck during the 'computeActualUsage' phase and increases the build time
of the 'buildHealth' task by as much as five times (5x).

For more info see: Stuck in computeActualUsageDebug for many minutes
#1186 (https://github.com/autonomousapps/
dependency-analysis-gradle-plugin/issues/1186)